Making search work

+++ Web/Shared/Models/PaginationResponse.cs
@@ -5,5 +5,5 @@ public class PaginationResponse<T>
     public List<T> Data { get; set; } = [];
     public int PageNumber { get; set; } = default!;
     public int PageSize { get; set; } = default!;
-    public int TotalPages { get; set; } = default!;
+    public long TotalPages { get; set; } = default!;
 }
